requirements.txt
# Token Forge CrewAI - Dépendances
crewai>=0.28.0
langchain-openai>=0.0.2
langchain-anthropic>=0.1.1
langchain-google-genai>=0.0.5
langchain-groq>=0.1.0
pyyaml>=6.0
python-dotenv>=1.0.0